+++ GreenBank/classes/Models/Payments.cs
@@ -10,14 +10,14 @@ namespace GreenBank.classes.Models
     {
         private int id;
         private DateTime payments_date;
-        private float amount;
+        private decimal amount;
         private string detail;
         private bool is_it_paid;
         private int loan_id;
 
         public int Id { get => id; set => id = value; }
         public DateTime Payments_date { get => payments_date; set => payments_date = value; }
-        public float Amount { get => amount; set => amount = value; }
+        public decimal Amount { get => amount; set => amount = value; }
         public string Detail { get => detail; set => detail = value; }
         public bool Is_it_paid { get => is_it_paid; set => is_it_paid = value; }
         public int Loan_id { get => loan_id; set => loan_id = value; }
